requirements.txt
python-dotenv
langchain
langchain-core
langchain-google-genai
langchain-community
pinecone
sentence-transformers